requirements.txt
Package                      Version
---------------------------- -----------
absl-py                      2.1.0
alembic                      1.15.1
annotated-types              0.7.0
asttokens                    3.0.0
astunparse                   1.6.3
betterproto                  2.0.0b6
certifi                      2025.1.31
cffi                         1.17.1
charset-normalizer           3.4.1
colorlog                     6.9.0
comm                         0.2.2
contourpy                    1.3.1
cryptography                 44.0.0
cycler                       0.12.1
debugpy                      1.8.13
decorator                    5.2.1
dnspython                    2.7.0
email_validator              2.2.0
executing                    2.2.0
flatbuffers                  25.2.10
fonttools                    4.56.0
gast                         0.6.0
google-pasta                 0.2.0
greenlet                     3.1.1
grpcio                       1.70.0
grpclib                      0.4.7
h2                           4.2.0
h5py                         3.13.0
hpack                        4.1.0
hyperframe                   6.1.0
id                           1.5.0
idna                         3.10
imbalanced-learn             0.13.0
imblearn                     0.0
in-toto-attestation          0.9.3
ipykernel                    6.29.5
ipython                      9.0.2
ipython_pygments_lexers      1.1.1
jedi                         0.19.2
joblib                       1.4.2
jupyter_client               8.6.3
jupyter_core                 5.7.2
kagglehub                    0.3.7
keras                        3.8.0
kiwisolver                   1.4.8
libclang                     18.1.1
Mako                         1.3.9
Markdown                     3.7
markdown-it-py               3.0.0
MarkupSafe                   3.0.2
matplotlib                   3.10.1
matplotlib-inline            0.1.7
mdurl                        0.1.2
ml-dtypes                    0.4.1
model-signing                0.2.0
multidict                    6.1.0
namex                        0.0.8
nest-asyncio                 1.6.0
numpy                        2.0.2
nvidia-cublas-cu12           12.5.3.2
nvidia-cuda-cupti-cu12       12.5.82
nvidia-cuda-nvcc-cu12        12.5.82
nvidia-cuda-nvrtc-cu12       12.5.82
nvidia-cuda-runtime-cu12     12.5.82
nvidia-cudnn-cu12            9.3.0.75
nvidia-cufft-cu12            11.2.3.61
nvidia-curand-cu12           10.3.6.82
nvidia-cusolver-cu12         11.6.3.83
nvidia-cusparse-cu12         12.5.1.3
nvidia-nccl-cu12             2.21.5
nvidia-nvjitlink-cu12        12.5.82
opt_einsum                   3.4.0
optree                       0.14.0
optuna                       4.2.1
packaging                    24.2
pandas                       2.2.3
parso                        0.8.4
pexpect                      4.9.0
pillow                       11.1.0
pip                          24.0
platformdirs                 4.3.6
prompt_toolkit               3.0.50
protobuf                     5.29.3
psutil                       7.0.0
ptyprocess                   0.7.0
pure_eval                    0.2.3
pyasn1                       0.6.1
pycparser                    2.22
pydantic                     2.10.6
pydantic_core                2.27.2
Pygments                     2.19.1
PyJWT                        2.10.1
pyOpenSSL                    25.0.0
pyparsing                    3.2.1
python-dateutil              2.9.0.post0
pytz                         2025.1
PyYAML                       6.0.2
pyzmq                        26.3.0
requests                     2.32.3
rfc3161-client               0.1.2
rfc8785                      0.1.4
rich                         13.9.4
scikit-learn                 1.6.1
scipy                        1.15.2
securesystemslib             1.2.0
setuptools                   65.5.0
sigstore                     3.6.1
sigstore-protobuf-specs      0.3.2
sigstore-rekor-types         0.0.18
six                          1.17.0
sklearn-compat               0.1.3
SQLAlchemy                   2.0.39
stack-data                   0.6.3
tensorboard                  2.18.0
tensorboard-data-server      0.7.2
tensorflow                   2.18.0
tensorflow-io-gcs-filesystem 0.37.1
termcolor                    2.5.0
threadpoolctl                3.5.0
tornado                      6.4.2
tqdm                         4.67.1
traitlets                    5.14.3
tuf                          5.1.0
typing_extensions            4.12.2
tzdata                       2025.1
urllib3                      2.3.0
wcwidth                      0.2.13
Werkzeug                     3.1.3
wheel                        0.45.1
wrapt                        1.17.2
